Split list-of-sounds message into chunks of <= 2000 characters
Discord messages cannot be longer than this limit. When we added enough
sounds to the bot, this limit was eventually reached while sending the
help message for !co-play.

The solution is to chunk the message into smaller sizes and send
multiple.

Also improve some of the help messages.

Also this reformats shit using PEP8 but I'm too lazy to remove the
whitespace changes.